Move gas and other helpers to separate files
